+++ src/index.css
@@ -1,1 +1,4 @@
 @import "tailwindcss";
+* {
+  font-family: 'Lexend Deca';
+}
